// jsonserver/types.ts
export interface Projectstype {
    id: string;
    projectName: string;
    projectdescription: string;
};

export interface V5datatype {
    id: string;
    name: string;
    category: string;
    partnum: string;
    type: string;
    datalocation: string;
    regtime: string;
    project: string;
};